test: move common fixtures to conftest files, organize structure of test files based on the template and cleanup the general structure of test suite for Pmodel
